add useSWRProvider
