+++ src/index.ts
@@ -1,5 +1,11 @@
 // Core APIs
-export { SWRConfig, mutate, createCache, unstable_serialize } from './use-swr'
+export {
+  SWRConfig,
+  mutate,
+  createCache,
+  useSWRProvider,
+  unstable_serialize
+} from './use-swr'
 
 // useSWR
 import useSWR from './use-swr'
