+++ src/use-swr.ts
@@ -1,25 +1,27 @@
-import { useCallback, useRef, useDebugValue } from 'react'
-import defaultConfig from './utils/config'
-import { wrapCache, SWRGlobalState, GlobalState } from './utils/cache'
+import { useCallback, useRef, useContext, useState, useDebugValue } from 'react'
+import { defaultConfig, defaultProvider } from './utils/config'
+import { wrapCache } from './utils/cache'
+import { SWRGlobalState, GlobalState } from './utils/global-state'
 import { IS_SERVER, rAF, useIsomorphicLayoutEffect } from './utils/env'
 import { serialize } from './utils/serialize'
 import { isUndefined, UNDEFINED } from './utils/helper'
-import ConfigProvider from './utils/config-context'
+import ConfigProvider, { SWRConfigContext } from './utils/config-context'
 import useStateWithDeps from './utils/state'
 import withArgs from './utils/resolve-args'
 import { subscribeCallback } from './utils/subscribe-key'
+import { broadcastState } from './utils/broadcast-state'
+import { getTimestamp } from './utils/timestamp'
+import { internalMutate } from './utils/mutate'
+
 import {
   State,
-  Broadcaster,
   Fetcher,
   Key,
-  MutatorCallback,
   SWRResponse,
   RevalidatorOptions,
   Configuration,
   SWRConfiguration,
   Cache,
-  ScopedMutator,
   SWRHook,
   StateUpdateCallback,
   RevalidateEvent,
@@ -27,128 +29,6 @@ import {
 } from './types'
 
 const WITH_DEDUPE = { dedupe: true }
-
-// Generate strictly increasing timestamps.
-let __timestamp = 0
-
-const broadcastState: Broadcaster = (
-  cache: Cache,
-  key,
-  data,
-  error,
-  isValidating,
-  shouldRevalidate = false
-) => {
-  const [EVENT_REVALIDATORS, STATE_UPDATERS] = SWRGlobalState.get(
-    cache
-  ) as GlobalState
-  const revalidators = EVENT_REVALIDATORS[key]
-  const updaters = STATE_UPDATERS[key]
-
-  // Always update states of all hooks.
-  if (updaters) {
-    for (let i = 0; i < updaters.length; ++i) {
-      updaters[i](data, error, isValidating)
-    }
-  }
-
-  // If we also need to revalidate, only do it for the first hook.
-  if (shouldRevalidate && revalidators && revalidators[0]) {
-    return revalidators[0](RevalidateEvent.MUTATE_EVENT).then(() =>
-      cache.get(key)
-    )
-  }
-
-  return Promise.resolve(cache.get(key))
-}
-
-const internalMutate = async <Data>(
-  cache: Cache,
-  _key: Key,
-  _data?: Data | Promise<Data | undefined> | MutatorCallback<Data>,
-  shouldRevalidate = true
-) => {
-  const [key, , keyErr] = serialize(_key)
-  if (!key) return UNDEFINED
-
-  const [, , MUTATION_TS, MUTATION_END_TS] = SWRGlobalState.get(
-    cache
-  ) as GlobalState
-
-  // if there is no new data to update, let's just revalidate the key
-  if (isUndefined(_data)) {
-    return broadcastState(
-      cache,
-      key,
-      cache.get(key),
-      cache.get(keyErr),
-      UNDEFINED,
-      shouldRevalidate
-    )
-  }
-
-  let data: any, error: unknown
-
-  // Update global timestamps.
-  const beforeMutationTs = (MUTATION_TS[key] = ++__timestamp)
-  MUTATION_END_TS[key] = 0
-
-  if (typeof _data === 'function') {
-    // `_data` is a function, call it passing current cache value.
-    try {
-      _data = (_data as MutatorCallback<Data>)(cache.get(key))
-    } catch (err) {
-      // if `_data` function throws an error synchronously, it shouldn't be cached
-      _data = UNDEFINED
-      error = err
-    }
-  }
-
-  // `_data` is a promise/thenable, resolve the final data first.
-  if (_data && typeof (_data as Promise<Data>).then === 'function') {
-    // This means that the mutation is async, we need to check timestamps to
-    // avoid race conditions.
-    try {
-      data = await _data
-    } catch (err) {
-      error = err
-    }
-
-    // Check if other mutations have occurred since we've started this mutation.
-    // If there's a race we don't update cache or broadcast the change,
-    // just return the data.
-    if (beforeMutationTs !== MUTATION_TS[key]) {
-      if (error) throw error
-      return data
-    }
-  } else {
-    data = _data
-  }
-
-  if (!isUndefined(data)) {
-    // update cached data
-    cache.set(key, data)
-  }
-  // Always update or reset the error.
-  cache.set(keyErr, error)
-
-  // Reset the timestamp to mark the mutation has ended
-  MUTATION_END_TS[key] = ++__timestamp
-
-  // Update existing SWR Hooks' internal states:
-  return broadcastState(
-    cache,
-    key,
-    data,
-    error,
-    UNDEFINED,
-    shouldRevalidate
-  ).then(res => {
-    // Throw error or return data
-    if (error) throw error
-    return res
-  })
-}
 
 export const useSWRHandler = <Data = any, Error = any>(
   _key: Key,
@@ -287,7 +167,7 @@ export const useSWRHandler = <Data = any, Error = any>(
           }
 
           CONCURRENT_PROMISES[key] = fn(...fnArgs)
-          CONCURRENT_PROMISES_TS[key] = startAt = ++__timestamp
+          CONCURRENT_PROMISES_TS[key] = startAt = getTimestamp()
 
           newData = await CONCURRENT_PROMISES[key]
 
@@ -617,20 +497,29 @@ export const SWRConfig = Object.defineProperty(ConfigProvider, 'default', {
   default: Configuration
 }
 
-export const mutate = internalMutate.bind(
-  UNDEFINED,
-  defaultConfig.cache
-) as ScopedMutator
+export const mutate = defaultProvider.mutate
 
 export const createCache = <Data = any>(
   provider: Cache,
   options?: Partial<ProviderOptions>
 ) => {
-  const cache = wrapCache<Data>(provider, options)
-  return {
-    cache,
-    mutate: internalMutate.bind(UNDEFINED, cache) as ScopedMutator<Data>
-  }
+  return wrapCache<Data>(provider, options)
+}
+
+export const useSWRProvider = <Data = any>(
+  initialize?: () => Cache,
+  options?: Partial<ProviderOptions>
+) => {
+  const cache = useContext(SWRConfigContext).cache || defaultConfig.cache
+  const [newProvider] = useState(
+    () => initialize && wrapCache<Data>(initialize(), options)
+  )
+
+  // If a new cache is created, return the new provider.
+  if (newProvider) return newProvider
+
+  // Return the current cache provider.
+  return { cache, mutate: (SWRGlobalState.get(cache) as GlobalState)[6] }
 }
 
 export const unstable_serialize = (key: Key) => serialize(key)[0]
