+++ src/utils/cache.ts
@@ -1,26 +1,16 @@
 import { provider as defaultProvider } from './web-preset'
 import { IS_SERVER } from './env'
 import { UNDEFINED } from './helper'
+import { internalMutate } from './mutate'
+import { SWRGlobalState } from './global-state'
 
 import {
   Cache,
+  ScopedMutator,
   RevalidateCallback,
-  StateUpdateCallback,
   ProviderOptions,
   RevalidateEvent
 } from '../types'
-
-export type GlobalState = [
-  Record<string, RevalidateCallback[]>, // EVENT_REVALIDATORS
-  Record<string, StateUpdateCallback[]>, // STATE_UPDATERS
-  Record<string, number>, // MUTATION_TS
-  Record<string, number>, // MUTATION_END_TS
-  Record<string, any>, // CONCURRENT_PROMISES
-  Record<string, number> // CONCURRENT_PROMISES_TS
-]
-
-// Global state used to deduplicate requests and store listeners
-export const SWRGlobalState = new WeakMap<Cache, GlobalState>()
 
 function revalidateAllKeys(
   revalidators: Record<string, RevalidateCallback[]>,
@@ -34,13 +24,16 @@ function revalidateAllKeys(
 export function wrapCache<Data = any>(
   provider: Cache<Data>,
   options?: Partial<ProviderOptions>
-): Cache<Data> {
+) {
   const EVENT_REVALIDATORS = {}
   const opts = { ...defaultProvider, ...options }
 
+  // Also create the mutate function that bound to the cache.
+  const mutate = internalMutate.bind(UNDEFINED, provider) as ScopedMutator<Data>
+
   // Initialize global state for the specific data storage that will be used to
   // deduplicate requests and store listeners.
-  SWRGlobalState.set(provider, [EVENT_REVALIDATORS, {}, {}, {}, {}, {}])
+  SWRGlobalState.set(provider, [EVENT_REVALIDATORS, {}, {}, {}, {}, {}, mutate])
 
   // Setup DOM events listeners for `focus` and `reconnect` actions.
   if (!IS_SERVER) {
@@ -63,5 +56,8 @@ export function wrapCache<Data = any>(
   // We might want to inject an extra layer on top of `provider` in the future,
   // such as key serialization, auto GC, etc.
   // For now, it's just a `Map` interface without any modifications.
-  return provider
+  return {
+    cache: provider,
+    mutate
+  }
 }
