+++ src/utils/timestamp.ts
@@ -1,0 +1,4 @@
+// Global timestamp.
+let __timestamp = 0
+
+export const getTimestamp = () => ++__timestamp
